Added version, some refacting

+++ src/test/java/skype2disk/mocks/BasePathMock.java
@@ -5,20 +5,27 @@ import java.io.IOException;
 
 import org.apache.commons.io.FileUtils;
 
+import com.google.inject.Singleton;
+
 import skype.BasePath;
 import testutils.IOHelper;
 
 public class BasePathMock implements BasePath {
 
-	private static final File CREATE_TEMP_DIR_OR_CRY = IOHelper.createTempDirOrCry();
+	private File CREATE_TEMP_DIR_OR_CRY;
 
 	@Override
 	public String getPath() {
+		if (CREATE_TEMP_DIR_OR_CRY == null)
+			CREATE_TEMP_DIR_OR_CRY = IOHelper.createTempDirOrCry();
 		return CREATE_TEMP_DIR_OR_CRY.getAbsolutePath();
 	}
 	
 	public void delete() {
 		try {
+			if (CREATE_TEMP_DIR_OR_CRY == null)
+				return;
+			
 			FileUtils.deleteDirectory(CREATE_TEMP_DIR_OR_CRY);
 		} catch (IOException e) {
 			throw new RuntimeException(e);
